refactor: extract APP-12 Rulebook into src/rulebook.js via a late-binding registry
Pure mechanical relocation of app.js's APP-12 chapter (RB_FOUNDATION, RB_TABS,
ruleOf, refPath, onInspectMove) into its own ES module. No behavior change.

This is the first extraction that needs something app.js itself still owns
(RULE_META, state). A plain `import {...} from '../app.js'` is UNSAFE here:
index.html loads the entry script as `app.js?v=<token>` (the cache-bust
query string), and ES modules cache by exact resolved URL — a bare `../app.js`
specifier from a sibling module resolves to a DIFFERENT URL, so the loader
silently evaluates a SECOND, fully independent copy of the whole app (its own
IDX, its own state, its own boot()) instead of sharing the real one. Confirmed
empirically: an earlier attempt to pull IDX/rentalUnits/etc. back into
src/derivations.js this way passed smoke (boots fine — nothing throws) but
failed logic-test 706/706, because IDX lookups silently read the phantom
copy's empty indexes instead of the real, live one. That attempt was reverted.

The fix used here instead: src/internals.js exports one shared object (APP),
populated by app.js (`Object.assign(APP, { RULE_META, state })`, placed right
before `boot()` so every chapter has already run) and read by rulebook.js at
CALL TIME via `APP.state`/`APP.RULE_META` (never at module-eval time). Every
module reaches internals.js through an ordinary relative specifier — it is
never anyone's <script> entry point, so there's exactly one instance, no
duplication trap. This pattern is documented in src/internals.js for reuse by
future extractions that need something still living in app.js.

Left a pointer banner at the old APP-12 location (box style, matching the
APP-01/02/23 precedent) so downstream APP-NN ids keep their position. Added
both new files to the code-map generator and the service worker's offline
SHELL precache.

Gates green: smoke, logic-test (706/706), lease/lease-deploy/promote/cachebust
suites, gen-rule-usage --check, check-window-catalog, gen-code-map --check.
